Add code for deduplicating feature names so they can fit in a single dataset or table. (#115)
This requires making the base Feature class an attrs dataclass, guaranteeing all implementation can have their names changed with attrs.evolve.

This code is not used anywhere yet; it should be used in the feature search runner, but I'm holding off on implementing it. The simple runner already has several problems, and I want to replace or at least heavily refactor it.