Make ProblemDescription serializable. (#334)
The type of `ProblemDescription.target_desired_outcome` was `ClassificationClass | RegressionDirection | None`, where `ClassificationClass = bool | int | str`. RegressionDirection was an StrEnum, which is a subclass of str, and serialized to json as a string; it was impossible to tell it apart from a string ClassificationClass.

The solution is use a polymorphic ADT instead.